+++ Code/EAI_3ModConfig.lua
@@ -1,0 +1,9 @@
+-- Code developed for Elevator A.I.
+-- Author @SkiRich
+-- All rights reserved, duplication and modification prohibited.
+-- You may not copy it, package it, or claim it as your own.
+-- Mod Created Sept 5th, 2018
+-- File Created Feb 21st, 2020
+-- Updated Feb 21st, 2020
+
+